+++ src/components/NsDrawer/index.ts
@@ -1,0 +1,1 @@
+export { default as NsDrawer } from './NsDrawer.vue'