+++ src/components/NsExpansionItem/index.ts
@@ -1,0 +1,1 @@
+export { default as NsExpansionItem } from './NsExpansionItem.vue'